refact: GraphSkeleton component
