+++ src/components/routes/(home)/components/GraphSkeleton/graphSkeleton.module.css
