+++ src/components/routes/(home)/components/GraphSkeleton/index.ts
